@@ -15,20 +15,20 @@
     "@skeletonlabs/skeleton": "2.7.0",
     "@skeletonlabs/tw-plugin": "0.3.1",
     "@sveltejs/adapter-auto": "^3.0.0",
-    "@sveltejs/kit": "^2.5.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tailwindcss/forms": "^0.5.7",
     "@types/node": "20.10.6",
     "autoprefixer": "10.4.16",
     "postcss": "8.4.32",
     "prisma": "^5.19.1",
     "supabase": "^1.192.5",
-    "svelte": "^4.2.7",
-    "svelte-check": "^3.6.0",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "tailwindcss": "3.4.0",
     "tslib": "^2.4.1",
-    "typescript": "^5.4.5",
-    "vite": "^5.0.0",
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4",
     "vite-plugin-tailwind-purgecss": "0.2.0",
     "vitest": "^2.0.5"
   },
@@ -50,7 +50,7 @@
     "prettier": "^3.2.5",
     "prettier-eslint": "^16.3.0",
     "svelte-plotly.js": "^0.4.0",
-    "vite": "^5.0.0",
+    "vite": "^5.4.4",
     "vite-plugin-tailwind-purgecss": "0.2.0"
   }
 }
