@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { createEventDispatcher } from 'svelte';
   import Icon from '@iconify/svelte';
-  export let muscleGroup = {muscleGroup: '', numSets: 0};
+  let { muscleGroup = $bindable({muscleGroup: '', numSets: 0}) } = $props();
 
   const dispatch = createEventDispatcher();
 </script>
@@ -25,7 +25,7 @@
 
 <input type='number' class='input' placeholder='Number of sets' bind:value={muscleGroup.numSets}/>  
 
-<button class='btn-icon variant-ghost-secondary' type='button' on:click={() => dispatch('remove')}><Icon icon="fa6-solid:minus" /></button>
+<button class='btn-icon variant-ghost-secondary' type='button' onclick={() => dispatch('remove')}><Icon icon="fa6-solid:minus" /></button>
 
 </div>
 
