@@ -8,8 +8,8 @@
 	import { storePopup } from '@skeletonlabs/skeleton';
 	storePopup.set({ computePosition, autoUpdate, offset, shift, flip, arrow });
 			
-	export let data
-	$: ({ supabase, session } = data)
+	let { data, children } = $props();
+	let { supabase, session } = $derived(data)
 
 	onMount(() => {
 		const { data } = supabase.auth.onAuthStateChange((_event, session) => {
@@ -46,22 +46,22 @@
 	<nav class="list-nav">
 		<ul>
 			<li>
-				<a class="btn btn-sm variant-ghost-primary" href="/account/view" on:click={closeDrawer}>Account</a>
+				<a class="btn btn-sm variant-ghost-primary" href="/account/view" onclick={closeDrawer}>Account</a>
 			</li>
 			<li>
-				<a class="btn btn-sm variant-ghost-primary" href="/landing" on:click={closeDrawer}>Home</a>
+				<a class="btn btn-sm variant-ghost-primary" href="/landing" onclick={closeDrawer}>Home</a>
 			</li>
 			<li>
-				<a class="btn btn-sm variant-ghost-primary" href="/exercises/create" on:click={closeDrawer}>Create Exercise</a>
+				<a class="btn btn-sm variant-ghost-primary" href="/exercises/create" onclick={closeDrawer}>Create Exercise</a>
 			</li>
 			<li>
-				<a class="btn btn-sm variant-ghost-primary" href="/exercises/list" on:click={closeDrawer}>Exercise List</a>
+				<a class="btn btn-sm variant-ghost-primary" href="/exercises/list" onclick={closeDrawer}>Exercise List</a>
 			</li>
 			<li>
-				<a class="btn btn-sm variant-ghost-primary" href="/programs/templateslist" on:click={closeDrawer}>Programs</a>
+				<a class="btn btn-sm variant-ghost-primary" href="/programs/templateslist" onclick={closeDrawer}>Programs</a>
 			</li>
 			<li>
-				<a class="btn btn-sm variant-ghost-primary" href="/account/workout-history" on:click={closeDrawer}>Workout History</a>
+				<a class="btn btn-sm variant-ghost-primary" href="/account/workout-history" onclick={closeDrawer}>Workout History</a>
 			</li>
 		</ul>
 
@@ -70,20 +70,26 @@
 
 </Drawer>
 <AppShell>
-	<svelte:fragment slot="header">
-		<!-- App Bar -->
-		<AppBar>
-			<svelte:fragment slot="lead">
-				{#if session}
-					<button class="text-xl font-extrabold uppercase" on:click={openDrawer}>Base45</button>
-				{:else}
-					<button class="text-xl font-extrabold uppercase" disabled>Base45</button>
-				{/if}
-			</svelte:fragment>
-			<svelte:fragment slot="trail">
-			</svelte:fragment>
-		</AppBar>
-	</svelte:fragment>
+	{#snippet header()}
+	
+			<!-- App Bar -->
+			<AppBar>
+				{#snippet lead()}
+					
+						{#if session}
+							<button class="text-xl font-extrabold uppercase" onclick={openDrawer}>Base45</button>
+						{:else}
+							<button class="text-xl font-extrabold uppercase" disabled>Base45</button>
+						{/if}
+					
+					{/snippet}
+				{#snippet trail()}
+					
+					
+					{/snippet}
+			</AppBar>
+		
+	{/snippet}
 	<!-- Page Route Content -->
-	<slot />
+	{@render children?.()}
 </AppShell>
