@@ -7,7 +7,11 @@
     import Icon, { iconLoaded } from '@iconify/svelte';
     import LinePlot from '$lib/components/LinePlot.svelte';
     
-    export let data: PageData;
+  interface Props {
+    data: PageData;
+  }
+
+  let { data }: Props = $props();
 
 </script>
 
@@ -16,7 +20,7 @@
 </svelte:head>
 
 {#if !data.profile}
-    <div class='placeholder animate-pulse'/>
+    <div class='placeholder animate-pulse'></div>
 {:else}
 
 <div class='card p-4 variant-glass-primary'>
@@ -24,45 +28,53 @@
     <section class='p-4'>
         <Accordion >
             <AccordionItem open>
-                <svelte:fragment slot="summary">
-                    Profile Details
-                </svelte:fragment>
-                <svelte:fragment slot="content">
-                    <div class='p-4'>
-                    <p>Display Name: {data.profile?.display_name}</p>
-                    <p>Date of Birth: {data.profile?.date_of_birth} </p>
-                    <p>Gender: {data.profile?.gender}</p>
-                    </div>
-                    <a href="/" class="btn btn-sm variant-ghost-secondary" data-sveltekit-preload-data="hover">Edit Profile</a>
+                {#snippet summary()}
+                  
+                      Profile Details
+                  
+                  {/snippet}
+                {#snippet content()}
+                  
+                      <div class='p-4'>
+                      <p>Display Name: {data.profile?.display_name}</p>
+                      <p>Date of Birth: {data.profile?.date_of_birth} </p>
+                      <p>Gender: {data.profile?.gender}</p>
+                      </div>
+                      <a href="/" class="btn btn-sm variant-ghost-secondary" data-sveltekit-preload-data="hover">Edit Profile</a>
 
-                </svelte:fragment>
+                  
+                  {/snippet}
             </AccordionItem>
             <AccordionItem>
-                <svelte:fragment slot='summary'>
-                    Weight History
-                </svelte:fragment>
-                <svelte:fragment slot='content'>
-                    {#if !data.weightHistoryData}
-                        <div class='placeholder'></div>
-                    {:else}
-                        <LinePlot data={data.weightHistoryData}/>
-                    {/if}
+                {#snippet summary()}
+                  
+                      Weight History
+                  
+                  {/snippet}
+                {#snippet content()}
+                  
+                      {#if !data.weightHistoryData}
+                          <div class='placeholder'></div>
+                      {:else}
+                          <LinePlot data={data.weightHistoryData}/>
+                      {/if}
 
-                    <form class='p-4' method="post" use:enhance action='?/addWeight'>
-                        <div class='input-group input-group-divider grid-cols-[1fr_1fr_3fr_1fr]'>
-                            <input class='input' name='value' type='number' step=0.1 placeholder="Weight">
-                            <select name='unit' value='lbs'>
-                                <option value='kg'>kg</option>
-                                <option value='lbs'>lbs</option>
-                            </select>
-                            <input name='date' type='date'>
-                            <button class='btn btn-sm variant-filled-primary' type='submit'>
-                                <Icon icon='fa6-solid:plus' />
-                            </button>
-                        </div>
+                      <form class='p-4' method="post" use:enhance action='?/addWeight'>
+                          <div class='input-group input-group-divider grid-cols-[1fr_1fr_3fr_1fr]'>
+                              <input class='input' name='value' type='number' step=0.1 placeholder="Weight">
+                              <select name='unit' value='lbs'>
+                                  <option value='kg'>kg</option>
+                                  <option value='lbs'>lbs</option>
+                              </select>
+                              <input name='date' type='date'>
+                              <button class='btn btn-sm variant-filled-primary' type='submit'>
+                                  <Icon icon='fa6-solid:plus' />
+                              </button>
+                          </div>
 
-                    </form>
-                </svelte:fragment>
+                      </form>
+                  
+                  {/snippet}
             </AccordionItem>
         </Accordion>
     </section>
