@@ -3,7 +3,11 @@
     import { Accordion, AccordionItem } from '@skeletonlabs/skeleton';
     import Icon from '@iconify/svelte';
     
-    export let data: PageData;
+    interface Props {
+        data: PageData;
+    }
+
+    let { data }: Props = $props();
 </script>
 
 <div class='card p-4'>
@@ -11,15 +15,21 @@
     <Accordion>
         {#each data.exerciseData as exercise}
             <AccordionItem>
-                <svelte:fragment slot="lead">
-                <Icon icon='fa6-solid:dumbbell'/>     
-                </svelte:fragment>
-                <svelte:fragment slot='summary'>
-                    {exercise.exercise_name}
-                </svelte:fragment>
-                <svelte:fragment slot='content'>
+                {#snippet lead()}
+                            
+                    <Icon icon='fa6-solid:dumbbell'/>     
+                    
+                            {/snippet}
+                {#snippet summary()}
+                            
+                        {exercise.exercise_name}
+                    
+                            {/snippet}
+                {#snippet content()}
+                            
 
-                </svelte:fragment>
+                    
+                            {/snippet}
             </AccordionItem>
         {/each}
     </Accordion>
